# openapi.yaml
openapi: 3.1.0

info:
  title: DealershipAI Action API
  version: 1.0.0
  description: >
    Action endpoints for Schema King GPT to retrieve, analyze, and (optionally) auto-deploy structured data
    optimizations across dealer sites.

servers:
  - url: https://api.schema.dealershipai.com
    description: Production server
  - url: https://dealershipai-app.com
    description: Current production deployment
  - url: https://dealership-ai-dashboard-rolrhzgi0-brian-kramer-dealershipai.vercel.app
    description: Latest Vercel deployment

paths:
  /api/ai-scores:
    get:
      operationId: getAIScores
      summary: Retrieve AIV / ATI / CRS metrics for a dealer
      description: Returns comprehensive AI visibility metrics including AIV (AI Visibility), ATI (Algorithmic Trust Index),
        and CRS (Composite Risk Score)
      parameters:
        - name: domain 
          in: query
          required: true
          schema:
            type: string
            pattern: ^https?://[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
          description: Dealership website URL
        - name: force_refresh
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Force refresh of cached scores
      responses:
        "200":
          description: AI scores retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    type: string
                  overall_score:
                    type: number
                    minimum: 0
                    maximum: 100
                  pillars:
                    type: object
                  revenue_impact:
                    type: object
                  action_items:
                    type: array
                    items:
                      type: string
        "400":
          description: Invalid domain format
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded

  /api/opportunities:
    get:
      operationId: listOpportunities
      summary: Fetch ranked schema-fix tasks by impact
      description: Returns a prioritized list of schema optimization opportunities ranked by potential impact
      parameters:
        - name: domain
          in: query
          required: true
          schema:
            type: string
            pattern: ^https?://[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
          description: Dealership website URL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum number of opportunities to return
      responses:
        "200":
          description: Opportunities retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  opportunities:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        description:
                          type: string
                        impact_score:
                          type: number
                        effort:
                          type: string
                          enum:
                            - low
                            - medium
                            - high
                        category:
                          type: string
                        estimated_aiv_gain:
                          type: number
        "400":
          description: Invalid domain format
        "401":
          description: Unauthorized

  /api/site-inject:
    post:
      operationId: runSchemaInject
      summary: Submit validated JSON-LD to dealer CMS
      description: Injects validated JSON-LD structured data into dealer website via CMS API or edge worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dealerId:
                  type: string
                  description: Dealer identifier
                jsonld:
                  type: string
                  description: Validated JSON-LD structured data (stringified)
                auto_fix:
                  type: boolean
                  default: false
                  description: Automatically apply fixes without approval
              required:
                - dealerId
                - jsonld
      responses:
        "200":
          description: Schema injected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                      - ok
                      - error
                  applied:
                    type: integer
                    description: Number of pages/sections where schema was applied
                  version_id:
                    type: string
                    description: Version ID for rollback capability
                  warnings:
                    type: array
                    items:
                      type: string
        "400":
          description: Invalid request body
        "401":
          description: Unauthorized
        "403":
          description: Domain ownership not verified
        "500":
          description: Internal server error

  /api/refresh:
    post:
      operationId: refreshDealerCrawl
      summary: Trigger a new scan and metric re-computation
      description: Forces a fresh crawl of the dealer website and recomputes all AI visibility metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: string
                  pattern: ^https?://[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
              required:
                - domain
      responses:
        "200":
          description: Refresh initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum:
                      - ok
                  message:
                    type: string
                  job_id:
                    type: string
                    description: Background job ID for tracking
        "400":
          description: Invalid domain format
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded

  /api/zero-click:
    get:
      operationId: checkZeroClick
      summary: Get AI Overview / Gemini inclusion rates per intent
      description: Returns zero-click inclusion rates across AI platforms (ChatGPT, Gemini, Perplexity) broken down by query intent
      parameters:
        - name: dealerId
          in: query
          required: true
          schema:
            type: string
          description: Dealer identifier
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Optional query intent filter (buy, sell, service, trade)
      responses:
        "200":
          description: Zero-click data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  dealer_id:
                    type: string
                  overall_coverage:
                    type: number
                    minimum: 0
                    maximum: 100
                    description: Overall zero-click coverage percentage
                  by_platform:
                    type: object
                    properties:
                      chatgpt:
                        type: number
                      gemini:
                        type: number
                      perplexity:
                        type: number
                  by_intent:
                    type: object
                    properties:
                      buy:
                        type: number
                      sell:
                        type: number
                      service:
                        type: number
                      trade:
                        type: number
        "400":
          description: Invalid dealerId
        "401":
          description: Unauthorized

  /api/ai/health:
    get:
      operationId: fetchAIHealth
      summary: Pull engine status and visibility trends for one dealer
      description: Returns comprehensive health metrics including engine status, visibility trends, and recent changes
      parameters:
        - name: dealerId
          in: query
          required: true
          schema:
            type: string
          description: Dealer identifier
        - name: intent
          in: query
          required: false
          schema:
            type: string
            enum:
              - buy
              - sell
              - service
              - trade
              - all
            default: all
          description: Filter by query intent
      responses:
        "200":
          description: Health data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  dealer_id:
                    type: string
                  status:
                    type: string
                    enum:
                      - healthy
                      - degraded
                      - critical
                  visibility_score:
                    type: number
                    minimum: 0
                    maximum: 100
                  trends:
                    type: object
                    properties:
                      last_7_days:
                        type: number
                        description: Change in visibility over last 7 days
                      last_30_days:
                        type: number
                        description: Change in visibility over last 30 days
                  engine_status:
                    type: object
                    properties:
                      crawler:
                        type: string
                        enum:
                          - operational
                          - degraded
                          - down
                      analyzer:
                        type: string
                        enum:
                          - operational
                          - degraded
                          - down
                      cache:
                        type: string
                        enum:
                          - operational
                          - degraded
                          - down
                  recent_alerts:
                    type: array
                    items:
                      type: object
        "400":
          description: Invalid dealerId
        "401":
          description: Unauthorized

  /api/analyze:
    post:
      operationId: analyzeDealership
      summary: Analyze dealership AI visibility (PLG Landing Page)
      description: Comprehensive AI visibility analysis for dealership domains with 5-pillar scoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: string
                url:
                  type: string
                  description: Alias for domain parameter
              required:
                - domain
      responses:
        "200":
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  overall:
                    type: number
                    minimum: 0
                    maximum: 100
                  aiVisibility:
                    type: number
                  zeroClick:
                    type: number
                  ugcHealth:
                    type: number
                  geoTrust:
                    type: number
                  sgpIntegrity:
                    type: number
                  competitorRank:
                    type: integer
                  totalCompetitors:
                    type: integer
                  revenueAtRisk:
                    type: number
                  domain:
                    type: string
        "400":
          description: Invalid domain format

  /api/share/track:
    post:
      operationId: trackShare
      summary: Track share event and unlock feature
      description: Records a share event and unlocks the specified feature for 24 hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain:
                  type: string
                featureName:
                  type: string
                platform:
                  type: string
                  enum:
                    - twitter
                    - linkedin
                    - facebook
                    - copy
                shareUrl:
                  type: string
                referralCode:
                  type: string
                sessionId:
                  type: string
              required:
                - featureName
                - platform
                - shareUrl
      responses:
        "200":
          description: Share tracked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  shareId:
                    type: string
                  unlockExpiresAt:
                    type: string
                    format: date-time
                  message:
                    type: string
        "400":
          description: Invalid request body
    get:
      operationId: checkUnlock
      summary: Check if feature is unlocked
      description: Verifies if a feature is currently unlocked for a user/session
      parameters:
        - name: featureName
          in: query
          required: true
          schema:
            type: string
        - name: domain
          in: query
          required: false
          schema:
            type: string
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Unlock status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  isUnlocked:
                    type: boolean
                  unlockExpiresAt:
                    type: string
                    format: date-time
                    nullable: true
                  timeRemaining:
                    type: integer
                    description: Seconds remaining until unlock expires